requirements.txt
discord.py
gspread
oauth2client
google-api-python-client